+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All changes to this framework is documented here in chronological order (latest
 first).
 
-- 1.0.22 **Under development**
+- 1.0.22
   - Upgrade to ZeroMQ 4.2.0 stable (up from 4.1.6).
   - Add `context.setBlocky` and `context.setMaxMessageSize` (new in 4.2)
   - Various documentation updates to manual and `SocketType`.